+++ src/components/MenuConfig.ts
@@ -41,11 +41,11 @@ import {
   GetRandomNodeFromSceneGraph,
 } from "../core/model/utils";
 import { processImageNodesInSceneGraph } from "../core/processors/imageBoxProcessor";
+import { DEMO_SCENE_GRAPHS, SceneGraphCategory } from "../data/DemoSceneGraphs";
 import {
   extractPositionsFromNodes,
   extractPositionsFromUserData,
 } from "../data/graphs/blobMesh";
-import { sceneGraphs } from "../data/graphs/sceneGraphLib";
 import { demoSongAnnotations } from "../mp3/data";
 import { demoSongAnnotations2 } from "../mp3/demoSongAnnotations247";
 import {
@@ -59,7 +59,6 @@ import {
   setLeftSidebarConfig,
   setRightSidebarConfig,
 } from "../store/workspaceConfigStore";
-import { SceneGraphCategory } from "./../data/graphs/sceneGraphLib";
 import { IMenuConfig, IMenuConfig as MenuConfigType } from "./UniAppToolbar";
 
 // const handleExportConfig = (sceneGraph: SceneGraph) => {
@@ -155,7 +154,7 @@ export class MenuConfig {
 
   private buildGraphMenu(): IMenuConfig {
     const graphMenu: IMenuConfig = {};
-    Object.entries(sceneGraphs).forEach(([_categoryKey, category]) => {
+    Object.entries(DEMO_SCENE_GRAPHS).forEach(([_categoryKey, category]) => {
       graphMenu[category.label] = {
         submenu: this.createGraphSubmenu(category),
       };
